Get tests passing on current versions
The restclient vs rest-client thing isn't a huge deal,
they both work. But since the gem is named wrong
(should be rest_client), and there is a gem that forked it
so some dude could do some custom tweak, and that dude
named his custom version of the gem correctly, it
is now super fkn easy to get the wrong rest-client,
as seen by its 1.2 million downloads. So, I just like to
be sure by being consistent, and as such, am requiring
the one that matches the gem name

The let -> def is b/c the :all blocks were depending
on the let blocks, but they run in different contexts,
and that's not a valid dependency (equivalent to a class
method calling an instance method).