Fix double escape bug, update Sinatra, don't literally start a server to test
Don't remember why I did that before
(though it did support someone cloning it in go using the same test suite).
Anyway, switch it to Rack::Test.

All the layers of abstraction, different tools trying to interpret the escape sequences in different ways,
was making it super hard to tell if it was escaped properly from the test.
So fuck it, get rid of that and just assert that the escaped verion is in the body...
turns out I was double escaping it, so remove the one in the template
as that offers the most flexibility (if your content is already escaped,
this will still work for you)